Improve Statisitic: date intervals, new image cat ^^, code improve and other

+++ app/src/main/java/com/egoriku/catsrunning/ui/statisticChart/FitChart.java
@@ -82,7 +82,7 @@ public class FitChart extends View {
     }
 
 
-    public void setMaxValue(float value) {
+    public void setMaxValue(int value) {
         maxValue = value;
     }
